[改善]複数のjsファイルで利用している関数式をモジュール化
